User jwt created

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (req) => {
+    await req.jwtVerify();
+  });
+
+  app.get("/memories", async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -19,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (req) => {
+  app.get("/memories/:id", async (req, res) => {
     // const { id } = req.params;
 
     const paramsSchema = z.object({
@@ -33,6 +40,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return res.status(401).send();
+    }
 
     return memory;
   });
@@ -51,14 +62,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "36a8680a-84d4-4894-8f08-5296093ba2f6",
+        userId: req.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (req) => {
+  app.put("/memories/:id", async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -73,7 +84,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(req.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== req.user.sub) {
+      return res.status(401).send;
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -87,12 +108,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (req) => {
+  app.delete("/memories/:id", async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(req.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== req.user.sub) {
+      return res.status(401).send;
+    }
 
     await prisma.memory.delete({
       where: {
